pyproject.toml
[project]
name = "ntrip-caster"
version = "2.2.0"
description = "High-performance NTRIP Caster with RTCM parsing"
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.3",
    "flask-socketio==5.3.6",
    "psutil==5.9.5",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyproj>=3.6.1",
    "pyrtcm>=1.1.0",
    "python-socketio==5.8.0",
    "pyyaml>=6.0.3",
    "reverse-geocode>=1.6.0",
    "setuptools>=78.1.1",
    "werkzeug>=3.0.6",
]

[project.scripts]
ntrip-caster = "ntrip_caster.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ntrip_caster"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = [
    "PLR0913",
    "PLR0912",
    "PLR0915",
    "PLR2004",
    "PLR0911",
    "PLC0415",
    "PLW0603",
    "E402",
    "UP036",
    "RUF059",
    "E501",
    "UP008",
    "B007",
    "PLR1714",
    "E722",
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy",
    "pre-commit",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "requests>=2.32.5",
    "ruff",
    "types-psutil",
    "types-pyyaml",
    "types-requests",
    "types-setuptools",
]
